+++ src/commands.ts
@@ -16,6 +16,9 @@ import { handleSearch } from "./search";
 import { encodeUri } from "./showAstDocument";
 import { applyFixAndSave, isRealFileEditor, replaceAll } from "./utils";
 import type { ViewResults } from "./webviews/types/results";
+import { setupMcp } from "./mcp";
+import fs from "fs";
+import path from "node:path";
 
 /*****************************************************************************/
 /* Prelude */
@@ -95,6 +98,38 @@ export function registerCommands(env: Environment): Disposable[] {
         } else if (resp == "Do not show again") {
           env.showNudges = false;
         }
+      }
+    }),
+
+    vscode.commands.registerCommand("semgrep.mcpSetup", async () => {
+      const repoPath = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath;
+      if (
+        // make sure the repo path is valid
+        !repoPath ||
+        // For now, only set up MCP for Cursor
+        vscode.env.uriScheme !== "cursor" ||
+        // and not if the user has already said no
+        env.foldersWithNoMcpSetupNudges.includes(repoPath) ||
+        // and not if the user seems to have set it up already
+        fs.existsSync(path.join(repoPath, ".cursor", "rules", "semgrep.mdc"))
+      ) {
+        return;
+      }
+      const resp = await vscode.window.showInformationMessage(
+        "Would you like to set up remote MCP scanning with Semgrep in this repository? [Learn more](https://mcp.semgrep.ai/)",
+        "Set up",
+        "Do not show again",
+      );
+      if (resp == "Set up") {
+        // TODO: handle multi-workspace case?
+        if (repoPath) {
+          await setupMcp(repoPath);
+        }
+      } else if (resp == "Do not show again") {
+        env.foldersWithNoMcpSetupNudges = [
+          ...env.foldersWithNoMcpSetupNudges,
+          repoPath,
+        ];
       }
     }),