+++ src/extension.ts
@@ -107,6 +107,7 @@ async function afterClientStart(context: ExtensionContext, env: Environment) {
       }
     }
   });
+  vscode.commands.executeCommand("semgrep.mcpSetup");
 }
 
 // Automatically invoked by VS Code's extension API